prodigi1-batch-service/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
pytz
pyjwt
python-jose
requests
httpx
prometheus_fastapi_instrumentator
sqlalchemy
pandas
openpyxl
psycopg2